requirements.txt
django==3.0.5
djangorestframework==3.11.0
markdown==3.2.1
django-filter==2.3.0
Pillow==7.0.0
psycopg2==2.8.5
djongo==1.3.3